# 백엔드/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HUFS Town API
  version: 0.5.0
  description: GDG HUFS SSO, spaces, invitations, map editing and world admission. Media signaling uses the authenticated world WebSocket contract; Node control routes are private. Never log credentials, invite codes or admission headers.
paths:
...
  /api/v1/spaces:
    get:
      operationId: listSpaces
      security: [{ sessionCookie: [] }]
      description: Server-side substring search and pagination over spaces readable by the caller. Up to 50 results per page. Member spaces are ordered first by last visit; private spaces require membership, and unlisted spaces are listed when the caller is a member, has a pending request, or saved them as a favorite. Access-blocked and archived spaces are omitted unless view=archived; that view returns only spaces owned by the caller.
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
        - name: view
          in: query
          schema:
            {
              type: string,
              enum: [browse, mine, favorites, archived],
              default: browse,
            }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100000, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        "200": { $ref: "#/components/responses/SpacePage" }
        default: { $ref: "#/components/responses/SpaceError" }
    post:
      operationId: createSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Maximum 20 spaces per owner. Select one of the campus square, study space, or meetup hall starter maps; every new space receives an independent editable map. Older clients that omit templateId retain the OFFICE default. Creation returns OWNER membership. approvalRequired is supported only for PUBLIC and UNLISTED spaces.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: "#/components/schemas/SpaceEdit" }
                - type: object
                  required: [capacity]
                  properties:
                    capacity: { type: integer, minimum: 1, maximum: 100 }
                    templateId:
                      {
                        type: string,
                        enum: [CAMPUS_SQUARE, STUDY_SPACE, MEETUP_HALL, OFFICE],
                        default: OFFICE,
                      }
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    get:
      operationId: getSpace
      security: [{ sessionCookie: [] }]
      description: PRIVATE requires membership; a denied read is indistinguishable from a missing space (404). UNLISTED can be read by ID by any active signed-in account.
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
    patch:
      operationId: updateSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Existing members retain access after visibility changes. Capacity and map template are immutable in this version.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpaceEdit" }
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}/archive:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    post:
      operationId: archiveSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Preserves the space and its data for restoration. Archived spaces disappear from normal lists and reject new access; an unexpired ownership transfer must be cancelled or resolved first.
      responses:
        "200": { $ref: "#/components/responses/SpaceArchiveState" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}/restore:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    post:
      operationId: restoreSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Restores a space from the owner's archived list and re-enables its preserved data and memberships.
      responses:
        "200": { $ref: "#/components/responses/SpaceArchiveState" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}/clone:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    post:
      operationId: cloneSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Creates a new PRIVATE space with the requested name, original description/capacity/template/email-domain policy, all maps and the current draft/published state. Map and custom-asset IDs are regenerated; only READY assets are copied. Approval and guest entry are disabled. The new space starts with only the caller as OWNER, no invitations or social/event records, and a fresh publication history. Maximum 20 owned spaces and 50 maps per space.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpaceCloneRequest" }
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}/favorite:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    put:
      operationId: favoriteSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Saves the caller's favorite. The caller must be able to read the space; PRIVATE spaces require membership and access blocks are enforced. Repeating the request is safe.
      responses:
        "200": { $ref: "#/components/responses/SpaceFavoriteState" }
        default: { $ref: "#/components/responses/SpaceError" }
    delete:
      operationId: unfavoriteSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Removes the caller's favorite. Safe to repeat and does not disclose whether the space is still readable.
      responses:
        "200": { $ref: "#/components/responses/SpaceFavoriteState" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}/admission:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    post:
      operationId: issueWorldAdmission
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Checks active account and space access, upserts membership and last visit. Nonmembers of approval-required spaces must first be approved or redeem a valid invitation. An optional mapId selects a published map in this space; omission selects its configured entry map. A reservationId enters the reserved private room from 15 minutes before its start through its end; reservationId cannot be combined with portal proof, and any supplied mapId must match the reservation map. Portal arrivals provide sourceSpaceId, sourceMapId, and portalId; the server verifies the published source portal and chooses its destination map and safe spawn. Returns a one-use 30-second Redis ticket bound to user/session/space/map and atomically reserves a seat against space-wide capacity. Supplying the current resumeToken reuses that participant's seat only when the token belongs to the same user and HTTP session. Joined seats are renewed as 20-second leases, including the 15-second reconnect grace. Connect with WebSocket protocols hufs-town-v2 and hufs-ticket.TICKET; server selects hufs-town-v2. Cookie and exact Origin are also required. Obtain a fresh ticket for every reconnect. Never put tickets in URLs or logs.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mapId:
                  {
                    type: string,
                    minLength: 1,
                    maxLength: 80,
                    pattern: "^[A-Za-z0-9_-]+$",
                  }
                sourceSpaceId:
                  {
                    type: string,
                    minLength: 1,
                    maxLength: 80,
                    pattern: "^[A-Za-z0-9_-]+$",
                  }
                sourceMapId:
                  {
                    type: string,
                    minLength: 1,
                    maxLength: 80,
                    pattern: "^[A-Za-z0-9_-]+$",
                  }
                portalId:
                  {
                    type: string,
                    minLength: 1,
                    maxLength: 80,
                    pattern: "^[A-Za-z0-9_-]+$",
                  }
                resumeToken:
                  {
                    type: string,
                    minLength: 1,
                    maxLength: 100,
                    pattern: "^[A-Za-z0-9-]+$",
                    description: Current world resume token; used to reclaim the same space seat.,
                  }
                reservationId:
                  {
                    type: string,
                    format: uuid,
                    description: Optional reservation ID to enter its private room during the allowed reservation window. Cannot be combined with portal proof.,
                  }
      responses:
        "200":
          description: Temporary admission.
          content:
            application/json:
              schema:
                type: object
                required: [ticket, expiresInSeconds]
                properties:
                  ticket: { type: string, minLength: 43, maxLength: 43 }
                  expiresInSeconds: { type: integer, const: 30 }
        default: { $ref: "#/components/responses/SpaceError" }
...
  /api/v1/spaces/{id}/members:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    get:
      operationId: listSpaceMembers
      security: [{ sessionCookie: [] }]
      description: OWNER or ADMIN. Lists current members; user IDs are used for management requests and are not rendered as profile data.
      responses:
        "200":
          description: Current space membership.
          content:
            application/json:
              schema:
                {
                  type: array,
                  items: { $ref: "#/components/schemas/SpaceMember" },
                }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/ownership-transfers/incoming:
    get:
      operationId: listIncomingOwnershipTransfers
      security: [{ sessionCookie: [] }]
      description: Returns unexpired ownership transfer requests addressed to the authenticated user, limited to spaces where they remain a member and the sender remains the owner.
      responses:
        "200":
          description: Pending requests that the caller can accept or decline.
          content:
            application/json:
              schema:
                type: array
                items:
                  { $ref: "#/components/schemas/IncomingOwnershipTransfer" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/invitations/incoming:
    get:
      operationId: listIncomingSpaceInvitations
      security: [{ sessionCookie: [] }]
      description: Lists active account-targeted space invitations for the authenticated user. Expired, revoked, fully used, archived, or already joined spaces are omitted.
      responses:
        "200":
          description: Invitations addressed to the caller.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/IncomingSpaceInvite" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/join-requests/incoming:
    get:
      operationId: listIncomingSpaceJoinRequests
      security: [{ sessionCookie: [] }]
      description: Lists up to 200 pending join requests across active, approval-required spaces where the authenticated user is the owner or an administrator. Space details are only returned to managers of those spaces.
      responses:
        "200":
          description: Pending requests for spaces the caller manages.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/IncomingSpaceJoinRequest" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/invitations/{inviteId}/accept:
    parameters:
      - name: inviteId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: acceptIncomingSpaceInvitation
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Accepts an unexpired account-targeted invitation addressed to the caller, subject to the space email-domain policy. Membership and invitation use count are updated atomically. Retrying after acceptance returns the current membership.
      responses:
        "200":
          description: The joined space.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/invitations/{inviteId}/decline:
    parameters:
      - name: inviteId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: declineIncomingSpaceInvitation
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Revokes an active account-targeted invitation addressed to the caller.
      responses:
        "200":
          description: Invitation declined.
          content:
            application/json:
              schema:
                type: object
                required: [declined]
                properties:
                  declined: { type: boolean, const: true }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}/ownership-transfer:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    get:
      operationId: getSpaceOwnershipTransfer
      security: [{ sessionCookie: [] }]
      description: OWNER only. Returns the current unexpired transfer request for this space, or an empty array when none exists.
      responses:
        "200":
          description: Current pending request.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SpaceOwnershipTransfer" }
        default: { $ref: "#/components/responses/SpaceError" }
    post:
      operationId: requestSpaceOwnershipTransfer
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Sends a seven-day request to a current member. Only one request may be pending per space. Ownership changes only after the recipient accepts; the previous owner remains an ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetUserId]
              properties:
                targetUserId: { type: string, format: uuid }
      responses:
        "200":
          description: Pending request created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SpaceOwnershipTransfer" }
        default: { $ref: "#/components/responses/SpaceError" }
    delete:
      operationId: cancelSpaceOwnershipTransfer
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Cancels the current pending request.
      responses:
        "200":
          description: Request cancelled.
          content:
            application/json:
              schema:
                type: object
                required: [cancelled]
                properties:
                  cancelled: { type: boolean, const: true }
        default: { $ref: "#/components/responses/SpaceError" }
  /api/v1/spaces/{id}/ownership-transfer/respond:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    post:
      operationId: respondToSpaceOwnershipTransfer
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: The addressed member accepts or declines. Acceptance atomically updates the space owner and member roles; the prior owner remains an ADMIN. Expired or stale requests are discarded without changing ownership.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [ACCEPT, DECLINE] }
      responses:
        "200":
          description: Request response result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OwnershipTransferResult" }
        default: { $ref: "#/components/responses/SpaceError" }
...
  /api/v1/spaces/{id}/access-blocks:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    get:
      operationId: listSpaceAccessBlocks
      security: [{ sessionCookie: [] }]
      description: OWNER or ADMIN. Lists up to 200 accounts currently denied entry to this space.
      responses:
        "200":
          description: Space access blocks.
          content:
            application/json:
              schema:
                {
                  type: array,
                  items: { $ref: "#/components/schemas/SpaceAccessBlock" },
                }
        default: { $ref: "#/components/responses/SpaceError" }
...
components:
...
  parameters:
    SpaceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: HUFS_TOWN_SESSION }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-TOKEN }
  schemas:
...
    SpaceEdit:
      type: object
      required: [name, description, visibility]
      properties:
        name:
          {
            type: string,
            minLength: 1,
            maxLength: 60,
            description: Trimmed; no control/format characters.,
          }
        description:
          {
            type: string,
            maxLength: 300,
            description: Trimmed single-line description; no control/format characters.,
          }
        visibility: { type: string, enum: [PUBLIC, UNLISTED, PRIVATE] }
        approvalRequired:
          {
            type: boolean,
            default: false,
            description: Only applies to PUBLIC or UNLISTED spaces. Existing members and valid invitees can enter without owner approval.,
          }
...
  responses:
...
    Space:
      description: Space metadata and the caller's role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Space" }
...
    SpacePage:
      description: Visible spaces for the requested server-side filter and page.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SpacePage" }
    SpaceError:
      description: 400 invalid input/invite/event date; 401 session missing; 403 owner/account/CSRF/approval/access-block denied; 404 missing or private space; 409 quota/request/event/ownership-transfer/archive conflict; 429 per-account rate or daily event limit; 503 storage unavailable. Uses the common code/message shape.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }
...